[Appinfo] Fix section_id being thrown out instead of being stored.
Section_ids are now stored with all the other parsed data, as we
need them during encoding the data back into VDF format.

Maintaining some kind of a table for the section names and corresponding
section ids was deemed unreliable. New sections may be added or the old
ones changed.